IOS Application
Free is an IOS application that allows one to connect with friends
based on the fact whether they are "Free" or not. The app restricts one
to call of photo/video message those who are busy at the moment. With
this app, one can make plans with friends who they know are free which
will save them a lot of time.